fix: combine examples as one; update docs
