--- backend/examples/helpers/inclusion_proof.rs
@@ -1,8 +1,0 @@
-use serde::Deserialize;
-
-#[derive(Debug, Deserialize)]
-pub struct InclusionProof {
-    pub leaf_hash: String,
-    pub root_hash: String,
-    pub proof: String,
-}
