--- backend/examples/helpers/mod.rs
@@ -1,1 +1,0 @@
-pub mod inclusion_proof;
