+++ zk_prover/src/circuits/utils.rs
@@ -17,7 +17,7 @@ use halo2_proofs::{
         kzg::{
             commitment::{KZGCommitmentScheme, ParamsKZG},
             multiopen::{ProverSHPLONK, VerifierSHPLONK},
-            strategy::{AccumulatorStrategy, SingleStrategy},
+            strategy::SingleStrategy,
         },
     },
     transcript::{
@@ -29,7 +29,7 @@ use regex_simple::Regex;
 use snark_verifier::{
     cost::CostEstimation,
     pcs::kzg::{Bdfg21, KzgAs},
-    system::halo2::{compile, transcript::evm::EvmTranscript, Config},
+    system::halo2::{compile, Config},
     verifier::plonk::PlonkSuccinctVerifier,
 };
 use snark_verifier_sdk::{evm::gen_evm_proof_shplonk, CircuitExt};
@@ -131,27 +131,6 @@ pub fn full_verifier(
         Blake2bRead<&[u8], G1Affine, Challenge255<G1Affine>>,
         SingleStrategy<'_, Bn256>,
     >(verifier_params, vk, strategy, instances, &mut transcript)
-    .is_ok()
-}
-
-/// Verifies a proof that generated by the `gen_evm_proof_shplonk` function.
-pub fn full_evm_verifier(
-    params: &ParamsKZG<Bn256>,
-    vk: &VerifyingKey<G1Affine>,
-    proof: Vec<u8>,
-    public_inputs: Vec<Vec<Fp>>,
-) -> bool {
-    let instance: Vec<&[Fp]> = public_inputs.iter().map(|input| &input[..]).collect();
-    let instances = &[&instance[..]];
-
-    let mut transcript = TranscriptReadBuffer::<_, G1Affine, _>::init(proof.as_slice());
-    verify_proof::<_, VerifierSHPLONK<'_, Bn256>, _, EvmTranscript<_, _, _, _>, _>(
-        params,
-        vk,
-        AccumulatorStrategy::new(params.verifier_params()),
-        instances,
-        &mut transcript,
-    )
     .is_ok()
 }
